requirements.txt
python-telegram-bot==20.0
requests==2.28.2
telepot==12.7
httpx==0.23.0